+++ src/services/nest.rs
@@ -1,10 +1,11 @@
 use crate::message::{Message, Reading, Type};
-use crate::threading;
+use crate::supervisor;
 use crate::value::Value;
 use crate::Result;
 use chrono::{DateTime, Local};
 use crossbeam_channel::Sender;
 use eventsource::reqwest::Client;
+use failure::format_err;
 use rouille::url::Url;
 use serde::Deserialize;
 use std::collections::HashMap;
@@ -22,17 +23,18 @@ pub fn spawn(service_id: &str, settings: &Settings, tx: &Sender<Message>) -> Res
     let settings = settings.clone();
     let tx = tx.clone();
 
-    threading::spawn(format!("my-iot::nest:{}", &service_id), move || {
+    supervisor::spawn(format!("my-iot::nest:{}", &service_id), move || -> Result<()> {
         let client = Client::new(Url::parse_with_params(URL, &[("auth", &settings.token)]).unwrap());
         for event in client {
             if let Ok(event) = event {
                 if let Some(event_type) = event.event_type {
                     if event_type == "put" {
-                        send_readings(&service_id, &serde_json::from_str(&event.data).unwrap(), &tx).unwrap();
+                        send_readings(&service_id, &serde_json::from_str(&event.data)?, &tx)?;
                     }
                 }
             }
         }
+        Err(format_err!("Event source client is unexpectedly exhausted"))
     })?;
 
     Ok(vec![])